requirements.txt
Django==1.8.5
django-grappelli==2.7.1
Pillow==3.0.0
sorl-thumbnail==12.3
tablib==0.10.0
wheel==0.24.0
